pass functions to actions
